requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.52.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sympy==1.12
tbb==2021.12.0
thop==0.1.1.post2209072238
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
ultralytics==8.2.22
urllib3==2.2.1
zipp==3.19.0
